+++ database/migrations/2021_02_20_042943_create_kasirs_table.php
@@ -17,8 +17,8 @@ class CreateKasirsTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->char('kode_user', 4)->unique();
-            $table->bigInteger('number_of_transaction')->default(0);
-            $table->decimal('total_penjualan')->nullable();
+            $table->integer('number_of_transaction')->default(0);
+            $table->bigInteger('total_penjualan')->nullable();
             $table->timestamps();
         });
     }
