+++ database/migrations/2021_02_20_092302_create_members_table.php
@@ -18,7 +18,7 @@ class CreateMembersTable extends Migration
             $table->string('kode_member', 11);
             $table->string('nama', 50);
             $table->string('no_telephone', 21)->unique();
-            $table->double('saldo', 12, 2)->nullable();
+            $table->bigInteger('saldo')->nullable();
             $table->timestamps();
         });
     }
